fix(scorecard): handle qodo review suggestions of entity owner fix
