+++ workspaces/scorecard/plugins/scorecard-backend/src/service/CatalogMetricService.ts
@@ -22,6 +22,7 @@ import {
   EntityMetricDetail,
 } from '@red-hat-developer-hub/backstage-plugin-scorecard-common';
 import { Entity } from '@backstage/catalog-model';
+import { normalizeOwnerRef } from '../utils/normalizeOwnerRef';
 import { MetricProvidersRegistry } from '../providers/MetricProvidersRegistry';
 import { NotFoundError, stringifyError } from '@backstage/errors';
 import {
@@ -359,7 +360,7 @@ export class CatalogMetricService {
         entityNamespace: entity.metadata.namespace,
         entityName: entity.metadata.name,
         entityKind: entity.kind,
-        owner: entity.spec?.owner as string,
+        owner: normalizeOwnerRef(entity.spec?.owner) ?? '',
         metricValue: row.value,
         timestamp: new Date(row.timestamp).toISOString(),
         status: row.status,
